Use unified structure for contrib imports
