+++ tests/contrib/pylons/test_pylons.py
@@ -1,4 +1,3 @@
-
 import time
 
 from nose.tools import eq_
